Extract CreateProjectCommand and prepare to move to remote api
